:octocat::snake: 3.3 hex, travis on osx, commands
This skips over printing things in hex if codecs raises a `lookupError`,
as printing to hex isn't the point of the test, it's just a nice thing
to do while testing to see what bits of the message ARE getting through.

I've changed the before-install stuff on travis to a simple block. Hopefully
it's the right syntax, I'm not on OSX.

Wrote more of the commands, I'll need to figure out how I want that
to be built and the best way I can think of is to just build them
and work on the documents at the same time.